Sebaszm/streamer face update

+++ Streamer/Administrator.cpp
@@ -114,7 +114,7 @@ namespace Player {
             auto it = _streamers.begin();
             for (; it != _streamers.end(); ++it) {
                 ASSERT((*it).second != nullptr);
-                if (((*it).second->Type() & streamType) != 0) {
+                if ((static_cast<uint32_t>((*it).second->Type()) & static_cast<uint32_t>(streamType)) != 0) {
                     IPlayerPlatform* player = (*it).second->Create();
                     if (player != nullptr) {
                         frontend = new Frontend(this, player);
